[bfops/cli-server-clear]: do thing
