+++ crates/cli/src/lib.rs
@@ -31,7 +31,6 @@ pub fn get_subcommands() -> Vec<Command> {
         dns::cli(),
         generate::cli(),
         list::cli(),
-        local::cli(),
         init::cli(),
         build::cli(),
         server::cli(),
@@ -56,7 +55,6 @@ pub async fn exec_subcommand(config: Config, cmd: &str, args: &ArgMatches) -> Re
         "dns" => dns::exec(config, args).await,
         "generate" => generate::exec(config, args),
         "list" => list::exec(config, args).await,
-        "local" => local::exec(config, args).await,
         "init" => init::exec(config, args).await,
         "build" => build::exec(config, args).await,
         "server" => server::exec(config, args).await,
